requirements.txt
urllib3
python-docx
